reorganize code so that the power pack management code and the apstra client code are moved out
