+++ SnowTickets/requirements.txt
@@ -1,3 +1,4 @@
 pysnow==0.7.17
 PyYAML==6.0.1
 urllib3==1.25.11
+Flask==3.1.0
